requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
python-dotenv==0.10.3
Werkzeug==0.16.0
